+++ src/main/java/controllers/album/ListAlbumController.java
@@ -29,6 +29,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import utils.InformationsUtils;
 
+import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
@@ -136,21 +137,29 @@ public class ListAlbumController implements Initializable {
 
     // ALBUM LISTENING ICON CLICKED
     public void albumListeningButtonClicked() {
+        AlbumDto albumSelected = ListAlbumSelectionListener.getAlbumSelected();
         boolean hasMusicWithNoFile = false;
+        boolean hasMusicWithWrongFile = false;
 
-        if (ListAlbumSelectionListener.getAlbumSelected() != null) {
+        if (albumSelected != null) {
             AlbumDao albumDao = new AlbumDao();
-            AlbumDb albumSelectedToDb = AlbumMapper.toDb(ListAlbumSelectionListener.getAlbumSelected());
+            AlbumDb albumSelectedToDb = AlbumMapper.toDb(albumSelected);
 
             hasMusicWithNoFile = albumDao.getMusiques(albumSelectedToDb)
                     .stream()
                     .anyMatch(musique -> musique.getNomFichierMusique() == null ||
                             "".equals(musique.getNomFichierMusique()));
+
+            hasMusicWithWrongFile = Boolean.TRUE.equals(albumDao.getMusiques(albumSelectedToDb)
+                    .stream()
+                    .map(musique -> !(new File(musique.getNomFichierMusique()).exists()))
+                    .reduce((wrongFile1, wrongFile2) -> wrongFile1 || wrongFile2)
+                    .orElse(Boolean.FALSE));
         }
 
-        if (ListAlbumSelectionListener.getAlbumSelected() == null) {
+        if (albumSelected == null) {
             this.showAlbumSelectionErrorPopUp();
-        } else if (hasMusicWithNoFile) {
+        } else if (hasMusicWithNoFile || hasMusicWithWrongFile) {
             MusicErrorModal.getMusicFileErrorAlert(TypeSource.ALBUM);
         } else {
             Stage stage = new Stage();